support map node for json extract
